+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-28T12:00:13.551Z")
+    created: Optional[datetime] = parse_datetime("2025-09-01T07:54:08.422Z")
     r"""The date and time the resource was created"""