+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-28T12:00:09.894Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-01T07:54:04.987Z")
     r"""The date and time the resource was last updated"""