+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K3R8SPE2ZZEN5VP69ABFX4PG"
+        "01K4249YQW1J6HC63APKZC3VGA"
     )
     r"""The id of the resource"""