+++ packages/orq-rc/RELEASES.md
@@ -2449,3 +2449,13 @@ Based on:
 - [python v3.12.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.21] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.21 - packages/orq-rc
+
+## 2025-09-01 07:55:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.9 (2.687.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.22 - packages/orq-rc